Task 5 completed -> all tasks completed
