+++ lab-2/src/main/java/by/bsu/dependency/example/usageExample/databases/NameDatabase.java
@@ -1,4 +1,4 @@
-package by.bsu.dependency.example.usage.databases;
+package by.bsu.dependency.example.usageExample.databases;
 
 import by.bsu.dependency.annotation.Bean;
 import by.bsu.dependency.annotation.BeanScope;
