+++ lab-2/src/main/java/by/bsu/dependency/example/usageExample/exception/NoEarsException.java
@@ -1,4 +1,4 @@
-package by.bsu.dependency.example.usage.exception;
+package by.bsu.dependency.example.usageExample.exception;
 
 public class NoEarsException extends RuntimeException {
     public NoEarsException(String message) {
