+++ lab-2/src/main/java/by/bsu/dependency/example/usageExample/humanHead/Nose.java
@@ -1,26 +1,24 @@
-package by.bsu.dependency.example.usage.humanHead;
+package by.bsu.dependency.example.usageExample.humanHead;
 
 import by.bsu.dependency.annotation.Bean;
 import by.bsu.dependency.annotation.BeanScope;
 import by.bsu.dependency.annotation.Inject;
 import by.bsu.dependency.annotation.PostConstruct;
-import by.bsu.dependency.example.usage.Randomizer;
+import by.bsu.dependency.example.usageExample.Randomizer;
+import lombok.Getter;
 
 @Bean(name="NOSE", scope= BeanScope.PROTOTYPE)
 public class Nose {
     @Inject
     Randomizer randomizer;
 
+    @Getter
     private int size;
 
     @PostConstruct
     private void init(Randomizer randomizer) {
         this.randomizer = randomizer;
         size = this.randomizer.nextInt(0, 100);
-    }
-
-    public int getSize() {
-        return size;
     }
 
     public String sniffSomething(String object) {
